+++ PrimeCalcFast.py
@@ -1,7 +1,8 @@
-import itertools
-import multiprocessing
 from math import sqrt
 from time import perf_counter
+
+calcTo: int = int(input("Calculate the nth prime:"))
+time1 = perf_counter()
 
 
 def prime(num: int) -> bool:
@@ -12,63 +13,31 @@ def prime(num: int) -> bool:
     :returns: True if num is prime, False otherwise
     :rtype: bool
     """
-    for div in range(5, int(sqrt(num)) + 1, 6):
-        if num % div == 0 or num % (div + 2) == 0:
+    for div in range(5, int(sqrt(num))+1, 6):
+        if num % div == 0 or num % (div+2) == 0:
             return False
     return True
 
 
-def work(chunk: list) -> list:
-    to_return = []
-    for num in chunk:
-        if prime(num):
-            to_return.append(num)
-    return to_return
+def main():
+    primes_found: int = 2
+    current_num: int = 1
+
+    while primes_found < calcTo:
+        current_num += 4
+        primes_found += prime(current_num)
+
+        if primes_found >= calcTo:
+            return current_num
+
+        current_num += 2
+        primes_found += prime(current_num)
+
+    return current_num
 
 
-def num_range():
-    last = 6
-    while True:
-        # Find the starting points for `6n - 1` and `6n + 1`
-        first_minus_1 = (last + 1) // 6 * 6 - 1
-        if first_minus_1 < last:
-            first_minus_1 += 6
-
-        first_plus_1 = (last - 1) // 6 * 6 + 1
-        if first_plus_1 < last:
-            first_plus_1 += 6
-
-        # Generate ranges
-        result = []
-        minus_iter = range(first_minus_1, last+chunk_size + 1, 6)
-        plus_iter = range(first_plus_1, last+chunk_size + 1, 6)
-
-        # Merge the ranges directly
-        result.extend(minus_iter)
-        result.extend(plus_iter)
-        last += chunk_size
-        yield result
-
-
-if __name__ == '__main__':
-    calcTo: int = int(input("Calculate the nth prime:"))
-    time1 = perf_counter()
-
-    max_processes = multiprocessing.cpu_count()
-    chunk_size = 1000
-    processes = []
-    nums_found = 3
-    chunks = num_range()
-    current_chunks = []
-
-    with multiprocessing.Pool(max_processes) as pool:
-        while True:
-            current_chunks = list(itertools.chain(*pool.map(work, [next(chunks) for i in range(max_processes)])))
-            if len(current_chunks) + nums_found < calcTo:
-                nums_found += len(current_chunks)
-            else:
-                current_chunks.sort()
-                time2 = perf_counter()
-                print(f'Your #: {current_chunks[calcTo%chunk_size*max_processes]}\nTime: {time2 - time1}sec')
-                print(chunk_size,max_processes,calcTo)
-                break
+if __name__ == "__main__":
+    num = main()
+    time2 = perf_counter()
+    print(f'Your #: {num} \
+          \nTime: {time2-time1}sec')
